requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-qdrant
qdrant-client
groq
fastembed
pypdf
pymupdf
python-dotenv
docxtpl